requirements.txt
numpy==1.26.4
dlib>=19.7.0
opencv-python==4.6.0.66
Pillow>=10.0.0
torch
torchvision
pandas>=1.0.0
scikit-image
joblib
PyQt5